1.feat:支持ip记录

+++ license-server-core/src/main/java/com/yozosoft/licenseserver/model/dto/CdKeyClientDTO.java
@@ -19,7 +19,7 @@ public class CdKeyClientDTO {
 
     private String mac;
 
-    private Byte ip;
+    private Integer ip;
 
     private Date expireDate;
